Move MaskBounds into the ClipSources (from the MaskCacheInfo struct).
