+++ webrender/src/render_task.rs
@@ -286,7 +286,7 @@ impl RenderTask {
                 return false;
             }
 
-            match clip_info.mask_cache_info.bounds.inner {
+            match clip_info.bounds.inner {
                 Some(ref inner) if !inner.device_rect.is_empty() => {
                     inner_rect = inner_rect.and_then(|r| r.intersection(&inner.device_rect));
                     !inner.device_rect.contains_rect(&task_rect)
